backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# LLM & Embeddings
openai==1.58.1
tiktoken==0.8.0
sentence-transformers==3.3.1

# Vector Store
chromadb==0.5.23

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Task Queue
celery[redis]==5.4.0
redis==5.2.1

# Storage
boto3==1.35.81

# Utilities
pydantic[email]==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Code Quality
black==24.10.0
flake8==7.1.1
